requirements.txt
PyHamcrest
pytest-bdd==8.1.0
pytest==8.3.4
requests
selenium==4.28.1